services/rapid-x-agent/requirements.txt
livekit-agents[deepgram,openai,silero,noise-cancellation]>=0.11
python-dotenv>=1.0.0
